refactor: test utils
